pyproject.toml
[tool.poetry]
name = "fast_api_ws_generator"
version = "0.1.0"
description = ""
authors = ["Gleb Tocarenco <gleb.tocarenco@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.63.0"}
uvicorn = {extras = ["standard"], version = "^0.13.4"}
websocket-client = "^0.58.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
